feat: separate table components for dashboard and user
