+++ fe/app/page.tsx
@@ -3,6 +3,7 @@ import { DataTable } from "@/components/data-table";
 import { SectionCards } from "@/components/section-cards";
 
 import data from "./dashboard/data.json";
+import { columns } from "./dashboard/column";
 
 export default function Page() {
   return (
@@ -11,7 +12,7 @@ export default function Page() {
       <div className="px-4 lg:px-6">
         <ChartAreaInteractive />
       </div>
-      <DataTable data={data} />
+      <DataTable columns={columns} data={data} />
     </>
   );
 }
